+++ Staging/output_sql/sa_rpu.sql
@@ -2,5 +2,5 @@ CREATE TABLE sa_rpu (
 	column1 INTEGER,
 	mois VARCHAR,
 	finess VARCHAR,
-	total INTEGER
+	total REAL
 )
